refactor: useBookThumbnail hook 추출
